Fix perm_invent messing up dumplog.
display_inventory(), and dump_inventory() both call display_pickinv(),
which unconditionally calls various dump_x() routines, which write
to the dump_logs if they are open.  This was ok until something in
perm_invent triggers an update during the dump process, which causes
the inventory to be written multiple times.  Attempted solution is
to add want_dump parameter to display_pickinv() and only call dump_x
routines if this flag is set.